+++ src/calendar/style.js
@@ -9,6 +9,13 @@ export default function getStyle(theme={}) {
     container: {
       backgroundColor: appStyle.calendarBackground
     },
+    dayContainer: {
+      flex: 1, 
+      alignItems: 'center'
+    },
+    emptyDayContainer: {
+      flex: 1
+    },
     monthView: {
       backgroundColor: appStyle.calendarBackground
     },